Modified nav-menu behavior to go to other page when the page requests it
It didn't make any sense that you get back to home after clicking on
the links to go to the open-source and privacy pages, because you had
to go through the legal page before reaching those.

This is a little tricky solution and I'm pretty sure that there's a
better way to achieve this, but it works so I'll leave it like it.